optimization code

+++ job/fetch.js
@@ -55,7 +55,7 @@ async function startCheck() {
   await bluebird.map(docs, async doc => {
     let id = doc.id;
     const value = String(doc.inData?.value);
-    if (value && !value.substring(value.length - 4).startsWith('90')) {
+    if (doc.inData && doc.inData.value && !value.substring(value.length - 4).startsWith('90')) {
       logger.info(`delete by value: ${value}, transcationId: ${doc.transcationId}`,)
       await makerTxModel.findOneAndDelete({id: id});
       return
